Prevent fork() from crashing after calls into InlineFortran
When a Python module is loaded from a DLL, it never gets unloaded during the lifetime of that Python process.  This means that when the process is forked (on Cygwin), in the process of setting up the child process all DLLs that were loaded in the parent need to be loaded in the child.  If the DLL itself was deleted it can't be loaded, and fork() fails.

By far the easiest solution is to not immediately delete temporary DLLs.  Instead they will be deleted when the process exits, by the atexit handler.  This is fine, just as long the DLLs stick around for the rest of the process's lifetime.

+++ src/sage/misc/inline_fortran.py
@@ -1,7 +1,11 @@
 """
 Fortran compiler
 """
-import os, imp, shutil
+
+import imp
+import os
+import shutil
+import sys
 
 from sage.misc.temporary_file import tmp_dir
 
@@ -133,11 +137,16 @@ class InlineFortran:
                 print(log_string)
         finally:
             os.chdir(old_cwd)
-            try:
-                shutil.rmtree(mytmpdir)
-            except OSError:
-                # This can fail for example over NFS
-                pass
+
+            if sys.platform != 'cygwin':
+                # Do not delete temporary DLLs on Cygwin; this will cause
+                # future forks of this process to fail.  Instead temporary DLLs
+                # will be cleaned up upon process exit
+                try:
+                    shutil.rmtree(mytmpdir)
+                except OSError:
+                    # This can fail for example over NFS
+                    pass
 
         for k, x in m.__dict__.iteritems():
             if k[0] != '_':
